refactor(zimbra): fix circular dependencies
ref: #MANAGER-18467
