+++ packages/manager/apps/zimbra/src/pages/dashboard/autoReplies/AutoReplies.types.ts
@@ -1,0 +1,10 @@
+import { ResourceStatus } from '@/data/api';
+
+export type AutoReplyItem = {
+  id: string;
+  name: string;
+  from: string;
+  until: string;
+  copyTo: string;
+  status: keyof typeof ResourceStatus;
+};
