+++ packages/manager/apps/zimbra/src/pages/dashboard/domains/Domains.types.ts
@@ -1,0 +1,11 @@
+import { ResourceStatus } from '@/data/api';
+
+export type DomainItem = {
+  id: string;
+  name: string;
+  organizationId: string;
+  organizationLabel: string;
+  account: number;
+  status: keyof typeof ResourceStatus;
+  cnameToCheck?: string;
+};
