+++ packages/manager/apps/zimbra/src/pages/dashboard/emailAccounts/settings/aliases/Aliases.types.ts
@@ -1,0 +1,7 @@
+import { ResourceStatus } from '@/data/api';
+
+export type AliasItem = {
+  id: string;
+  alias: string;
+  status: keyof typeof ResourceStatus;
+};
