+++ packages/manager/apps/zimbra/src/pages/dashboard/organizations/Organizations.page.tsx
@@ -18,7 +18,6 @@ import {
   PageLocation,
   useOvhTracking,
 } from '@ovh-ux/manager-react-shell-client';
-import { ResourceStatus } from '@/data/api';
 import { useGenerateUrl, useDebouncedValue, useOverridePage } from '@/hooks';
 import { useOrganizations, usePlatform } from '@/data/hooks';
 import ActionButton from './ActionButton.component';
@@ -26,14 +25,7 @@ import { BadgeStatus, LabelChip, IdLink } from '@/components';
 import { IAM_ACTIONS } from '@/utils/iamAction.constants';
 import { DATAGRID_REFRESH_INTERVAL, DATAGRID_REFRESH_ON_MOUNT } from '@/utils';
 import { ADD_ORGANIZATION } from '@/tracking.constants';
-
-export type OrganizationItem = {
-  id: string;
-  name: string;
-  label: string;
-  account: number;
-  status: keyof typeof ResourceStatus;
-};
+import { OrganizationItem } from './Organizations.types';
 
 const columns: DatagridColumn<OrganizationItem>[] = [
   {
