+++ packages/manager/apps/zimbra/src/pages/dashboard/organizations/Organizations.types.ts
@@ -1,0 +1,9 @@
+import { ResourceStatus } from '@/data/api';
+
+export type OrganizationItem = {
+  id: string;
+  name: string;
+  label: string;
+  account: number;
+  status: keyof typeof ResourceStatus;
+};
