+++ packages/manager/apps/zimbra/src/pages/dashboard/redirections/Redirections.types.ts
@@ -1,0 +1,9 @@
+import { ResourceStatus } from '@/data/api';
+
+export type RedirectionItem = {
+  id: string;
+  from: string;
+  to: string;
+  organization: string;
+  status: keyof typeof ResourceStatus;
+};
